This includes some initial work on ToC subcommand for ADR and RFD.
I've also started work on linking ADRs and RFDs and was initially going
to use Comrak to create an AST so that I could find the position in the
AST to add the necessary links and insert additional nodes however
Comrak doesnt support Frontmatter which is a big blocker. I've also
found that I dont love the API. As a result I'm working on my own Markdown
parser to build an AST. So while i'm leaving Comrak in this commit it
will be removed.