Start Work on Build Command
Starting preliminary work on bulid command. This commit is mostly with
me experimenting with how to support zero configuration builds. Zero
configuration builds are how Doctavious attempts to determine what
framework is being used and how to build it without the user explicitly
providing build options in the doctavious yaml. This is needed to
support our desire to get users up and running as quickly as possible.