Implement Default For Frameworks
I wanted to have const or static instances for the frameworks but
running into issues with const/static limitations specifically
dynamically sized arrays. Info currently only has one, configs, which I
could move to the last field however we'll have at least one more for
framework detection, although that might be in a substruct, which i'm
not sure is allowed. As a result, we implemented Default and we'll just
have a vec we'll construct when we need it. This is only during `build`
command so creating the vec as needed is fine.