hack: add fn new to each framework
This commit adds `fn new(configs: Option<Vec<&'static str>>) -> Self`
for each framework. This allows us to override configs for sole purpose
of testing determining output directory from configs.

I'm not a fan of this type of change as it has no real value but it is
what it is and I dont have a better idea at the moment